Move the run detail pane into a panel, which magically makes the border layout work again on the runs page in all browsers. This is a terrific example of just how horrifyingly fucking broken ExtJS's layout management is: adding an extraneous element which has absolutely no semantic or even layout purpose according to the documentation just happens to make everything get along again. The only conclusion I can draw from this is that the ExtJS developers have access to considerably better drugs than I do or are gibbering morons on a level not seen since the American people decided to elect George W. Bush for a second time.
Of course, I had to get the border layout working, because the others are even
worse. The hbox layout doesn't have any way to set the actual height of the
managed layout, which includes no way to just say "fill the whole fucking
space, you idiot", which is what you'd actually want in 95% of cases.

In short, a message to ExtJS developers: your Javascript framework is fucking
bullshit, and is hands-down the worst piece of Javascript I've seen since the
language was renamed in the 90s.

*deep breath*

OK, so this should get us into a position to do a release, barring any further
ExtJS fuck-ups.