MAS-7: mdatasync_client: better rsyncing
This change adds recording of which files were updated in the
rsync. If a file is rsynced and doesn't need to be changed then it can
be marked as complete.

Also changed the function queue so that the callback function can be
called with parameters.

File archiving is made simpler.

Some logging is cleaned up.