MAS-8: client: try to fix unicode filename issues
The datasync server outputs utf-8 encoded json. So this data needs to
be put into Python unicode objects.

Any "casting" of string-like objects with str() is removed unless it
is guaranteed that they won't contain non-ascii characters.

Filenames put into and got out of rsync are encoded/decoded in the
system's preferred unicode encoding. On Linux this is usually utf-8,
on Windows it is possibly utf-16.