client: fix problems with rsync on windows
I made some mistakes in the fix for MAS-7 which appear on windows.

First is a use of os.path.join when the path name is expected to have
forward slashes.

Second is if rsync produced any error (e.g. could not create directory
'/home/Rodney/.ssh') then the check rsync call wouldn't be made, so
files wouldn't be marked complete.