MAS-16: ExtJS: Make it possible to use the rule generator start block
The panel sizing is still not ideal -- i.e. the height fixed and
doesn't fill the window. However many windows are like that in
Mastr-MS, and I don't have a clue how to make it right.

This change makes it possible to at least access the grid panel which
was previously obscured by buttons.