MAS-22: Merge MAUser into User, start trying to simplify code
Just added all the MAUser methods/properties into User and substitued
User with MAUser.

Removed some of the caching stuff, but not all yet. I don't think a
few extra db queries for user checks will drastically affect
performance.

Next step is to merge in the DBUserManager with User and simplify.

Then a migration to merge or rename some of the extra User fields/