MAS-22: Adjust old migrations so they work from a fresh start
If you change the user model in settings, django will never create
auth_user because it's not needed.

So we need to update previously run migrations so that they work with
users_user instead of auth_user.

Then a syncdb && migrate will work from an empty database.

Most of this commit is the work of sed, but there is also some new
trickiness in the user table rename migration.

The order of apps in settings was changed to but mastrms.users first.

The initial data json needed to be updated.

Also I needed to add an initial migration to create the users_user
table.