MAS-22: Unbreak "My Account" view
The javascript code expects the user data dict to be wrapped in a
list, so fixed that.

Made it so the user can only access his own details (this bug existed
before the MAS-22 changes).

Refactored the setting of the "organisation" attribute of the user.

Made the difference between /admin/userload and /users/userload a
little clearer.