MAS-36: Change InstrumentMethod.template to be a simple option
The template field used to be a string which was fed into mako.

Subsequent changes made the template field irrelevant and a fixed csv
template was used.

This commit changes template into an switch where there is currently a
single option -- csv.

It also changes the view to use the csv module to generate the
worklist (helpful, because it will escape any data which contains
commas).

It also fixes the model so it appears well in the admin interface.