rpm: Install app into a virtualenv with specific python version
In the spec file, the %{pyver} and %{pybasever} macros control which
python is used.

During the build process, a virtualenv is created with python packages
from the app's requirements.txt.

App configuration had to change to support virtualenv. The
defaultsettings and appsettings modules are removed from the rpm, but
/etc/ccgapps/appsettings/mastrms.py is still used if it exists.

The RPM depends on a python27-psycopg2 package which I have backported
from fedora. The other dependency, python27-mod_wsgi, is included in the
IUS repo.

It seems like the manage and WSGI scripts are slightly simpler after
this change.