setup.py: Move requirements out of text file, back into setup.py
At present using requirements.txt may cause confusion to developers.

The deprecated dependency_links option in setup.py will surely be
supported for a year or two, so we will keep using it for now.

This change also starts using the extra_requires option to allow dev and
test requirements to be expressed.