Production install config file rearrangement
Now the production settings are in:

  /etc/mastrms/mastrms.conf - key-value file containing common config
  /etc/mastrms/settings.py - django settings file

When the rpm is installed, the mastrms.prodsettings module is setup as a
symlink to /etc/mastrms/settings.py.

The manage and wsgi scripts load the conf file and put its values into
process environment variables. mastrms.prodsettings picks up these
variables and uses them.

The motivation for this change was to make configuration easier, both
for the sysadmin, and for installing configuration through puppet.

The idea is more or less like the django-twoscoops project:
  https://github.com/twoscoops/django-twoscoops-project