client: move files to top level, get rid of supportwin32
The presence of large files in supportwin32 is slowing down checkouts
and rpm builds. Most of them are easily obtainable, so I replaced them
with a shell script which downloads them.

Also I have moved the datasync client code to the top level where it is
easier to find.

There will be further changes required to fix up the client build
process.