rpm: unreverse static symlink, nuke staticfiles before unpacking rpm
I changed /usr/local/webapps/mastrms/static into a symlink but it's
better the other way around -- to have a symlink within the python
sources pointing to the known directory name.

This has caused slight problems with unpacking the RPM.

According to http://www.ibm.com/developerworks/library/l-rpm2/,
the order of operations during an upgrade is:

1. Run the %pre section of the RPM being installed.
2. Install the files that the RPM provides.
3. Run the %post section of the RPM.
4. Run the %preun of the old package.
5. Delete any old files not overwritten by the newer version. (This step
   deletes files that the new package does not require.)
6. Run the %postun hook of the old package.

So if I want to change the current staticfiles directory into a symlink,
the rpm needs to get rid of its contents first.

This page also shows possible values of %pre and %preun scriptlet
arguments:

http://fedoraproject.org/wiki/Packaging:ScriptletSnippets#Syntax